renames a-frame to aframe and aFrame to aframe
