+++ src/aframe/component/gblock.js
