+++ src/aframe/component/gblock/fragment-placeholder.glsl
