+++ src/aframe/component/gblock/fragment.glsl
