+++ src/aframe/component/gblock/vertex-placeholder.glsl
