+++ src/aframe/component/gblock/vertex.glsl
