+++ src/aframe/component/lighting.js
