+++ src/aframe/component/tour.js
