+++ src/aframe/three/data3d-view.js
