+++ src/aframe/three/data3d-view/io3d-material/vertex.glsl
