+++ src/aframe/three/data3d-view/set-material.js
