+++ src/aframe/three/data3d-view/set-material/fetch-texture/fetch-image-texture.js
