+++ src/aframe/three/data3d-view/set-material/load-texture-set.js
