+++ src/aframe/three/data3d-view/wireframe.js
