+++ src/io3d.js
@@ -4,7 +4,7 @@ import './core/bootstrap.js'
 import runtime from './core/runtime.js'
 import configs from './core/configs.js'
 
-import aFrame from './a-frame.js'
+import aframe from './aframe.js'
 import furniture from './furniture.js'
 import staging from './staging.js'
 import storage from './storage.js'
@@ -18,7 +18,8 @@ import utils from './utils.js'
 var io3d = {
 
   // APIs
-  aFrame: aFrame,
+  aFrame: aframe, // alias for legacy support
+  aframe: aframe,
   furniture: furniture,
   staging:staging,
   storage: storage,
